color_picker/requirements.txt
streamlit==1.1.0
pandas==1.3.1
numpy==1.19.5
scipy==1.7.0
webcolors==1.11.1
pandas_gbq==0.15.0
matplotlib==3.4.2
opencv-python==4.5.3.56
scikit-learn==0.24.2
detecto==1.2.1
pillow==8.3.1